Use tricky (NaN) values in many more RP tests
All of these pass today, either because they were already working (eg,
they use memcpy under the hood), or because of fixes that have already
landed. However: some of these will break when we introduce the scalar
promotion helpers (`F_`, etc).

Along the way, I updated various context structs to use integers rather
than floats for ambiguously typed data.

For now, this just gets these tests running.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/801156
Commit-Queue: Brian Osman <brianosman@google.com>

+++ src/core/SkRasterPipelineOpContexts.h
@@ -197,13 +197,13 @@ struct SkRasterPipeline_InitLaneMasksCtx {
 };
 
 struct SkRasterPipeline_ConstantCtx {
-    float value;
+    int32_t value;
     SkRPOffset dst;
 };
 
 struct SkRasterPipeline_UniformCtx {
-    float* dst;
-    const float* src;
+    int32_t* dst;
+    const int32_t* src;
 };
 
 struct SkRasterPipeline_BinaryOpCtx {
@@ -231,7 +231,7 @@ struct SkRasterPipeline_SwizzleCtx {
 };
 
 struct SkRasterPipeline_ShuffleCtx {
-    float* ptr;
+    int32_t* ptr;
     int count;
     uint16_t offsets[16];  // values must be byte offsets (4 * highp-stride * component-index)
 };
